+++ src/functions/drawSvgPath.ts
