+++ src/functions/generatePath.ts
