+++ src/functions/getBoundingBoxes.ts
