+++ src/functions/guaranteeWalkablePath.ts
