+++ src/stories/TestGraph.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactFlow from 'react-flow-renderer'
-import { SmartEdge, SmartEdgeProvider } from '../../index'
-import type { SmartEdgeOptions } from '../../index'
+import { SmartEdge, SmartEdgeProvider } from '../index'
+import type { SmartEdgeOptions } from '../index'
 import type { ReactFlowProps } from 'react-flow-renderer'
 
 export type GraphProps = ReactFlowProps
